e_devcrypto: default to not use digests in engine
Digests are almost always slower when using /dev/crypto because of the
cost of the context switches.  Only for large blocks it is worth it.

Also, when forking, the open context structures are duplicated, but the
internal kernel sessions are still shared between forks, which means an
update/close operation in one fork affects all processes using that
session.

This affects digests, especially for HMAC, where the session with the
key hash is used as a source for subsequent operations.  At least one
popular application does this across a fork.  Disabling digests by
default will mitigate the problem, while still allowing the user to
turn them on if it is safe and fast enough.

Signed-off-by: Eneas U de Queiroz <cote2004-github@yahoo.com>